src/requirements.txt
jsonrpclib-pelix
pymongo
pika
requests
redis
sklearn
numpy
scipy
python-dateutil
newspaper3k
tensorflow
pandas
watchdog